+++ packages/sdk/src/models/abstract_model_with_async_data.ts
@@ -30,6 +30,18 @@ class AbstractModelWithAsyncData<DataType, WatchableKey extends string> extends 
     _dataRetainCount: number;
     /** @internal */
     _unloadDataTimeoutId: null | TimeoutId;
+    /**
+     * This flag is used to keep track of models that have been
+     * forced to unload (regardless of the retain count). The force
+     * unload happens via _forceUnload method and the only proper use
+     * of that function is when the underlying data gets deleted while
+     * the model is still active. e.g. when a table is deleted in the
+     * main app while an instance of record_store is still alive.
+     * NOTE: Once set to true, it never goes back to false.
+     *
+     * @internal
+     */
+    _isForceUnloaded: boolean = false;
     /** @hidden */
     constructor(baseData: BaseData, modelId: string) {
         super(baseData, modelId);
@@ -51,6 +63,7 @@ class AbstractModelWithAsyncData<DataType, WatchableKey extends string> extends 
         callback: FlowAnyFunction,
         context?: FlowAnyObject | null,
     ): Array<WatchableKey> {
+        this._assertNotForceUnloaded();
         const validKeys = super.watch(keys, callback, context);
         for (const key of validKeys) {
             if (
@@ -88,8 +101,18 @@ class AbstractModelWithAsyncData<DataType, WatchableKey extends string> extends 
         }
         return validKeys;
     }
+    /** @inheritdoc */
+    get isDeleted(): boolean {
+        if (this._isForceUnloaded) {
+            return true;
+        }
+        return super.isDeleted;
+    }
     /** */
     get isDataLoaded(): boolean {
+        if (this.isDeleted) {
+            return false;
+        }
         return this._isDataLoaded;
     }
     /** @internal */
@@ -120,6 +143,7 @@ class AbstractModelWithAsyncData<DataType, WatchableKey extends string> extends 
      * Returns a Promise that will resolve once the data is loaded.
      */
     async loadDataAsync() {
+        this._assertNotForceUnloaded();
         if (this._unloadDataTimeoutId !== null) {
             // If we set a timeout to unload data, clear it since we are incrementing
             // the retain count and loading data.
@@ -155,6 +179,9 @@ class AbstractModelWithAsyncData<DataType, WatchableKey extends string> extends 
     // IMPORTANT: always call super.unloadData() from your override.
     /** */
     unloadData() {
+        if (this._isForceUnloaded) {
+            return;
+        }
         this._dataRetainCount--;
 
         if (this._dataRetainCount < 0) {
@@ -178,6 +205,17 @@ class AbstractModelWithAsyncData<DataType, WatchableKey extends string> extends 
                 this._onChangeIsDataLoaded();
             }, AbstractModelWithAsyncData.__DATA_UNLOAD_DELAY_MS);
         }
+    }
+
+    _forceUnload() {
+        while (this._dataRetainCount > 0) {
+            this.unloadData();
+        }
+        this._isForceUnloaded = true;
+    }
+
+    _assertNotForceUnloaded() {
+        invariant(!this._isForceUnloaded, 'model (%s) permanently deleted', this.id);
     }
 }