+++ packages/sdk/CHANGELOG.md
@@ -13,6 +13,8 @@ commit that affects the code in a way that consumers might care about should inc
 
 -   Fixed a bug that prevented TableOrViewQueryResult from notifying watchers about the
     creation/deletion of sorted fields.
+-   Fixed a bug where deleting and undeleting a table that was already loaded caused the App to
+    crash.
 
 ## [1.2.0](https://github.com/airtable/blocks/compare/@airtable/blocks@1.1.0...@airtable/blocks@1.2.0) - 2020-10-23